server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
asgi-lifespan==2.0.0
    # via -r requirements.in
attrs==22.2.0
    # via pytest
autoflake==2.0.0
    # via -r requirements.in
black==22.12.0
    # via -r requirements.in
bump2version==1.0.1
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   uvicorn
coverage[toml]==7.0.4
    # via pytest-cov
dnspython==2.2.1
    # via pymongo
docker==6.0.1
    # via -r requirements.in
exceptiongroup==1.1.0
    # via pytest
fastapi==0.89.0
    # via -r requirements.in
flake8==6.0.0
    # via -r requirements.in
flake8-fixme==1.1.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-metadata==6.0.0
    # via pytest-randomly
iniconfig==2.0.0
    # via pytest
isort==5.11.4
    # via -r requirements.in
loguru==0.6.0
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mimesis==4.1.3
    # via
    #   -r requirements.in
    #   pytest-mimesis
motor==3.1.1
    # via -r requirements.in
mypy==0.991
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==23.0
    # via
    #   docker
    #   pytest
pathspec==0.10.3
    # via black
platformdirs==2.6.2
    # via black
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pydantic==1.10.4
    # via fastapi
pyflakes==3.0.1
    # via
    #   autoflake
    #   flake8
pymongo==4.3.3
    # via motor
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-env
    #   pytest-mimesis
    #   pytest-randomly
pytest-asyncio==0.20.3
    # via -r requirements.in
pytest-cov==4.0.0
    # via -r requirements.in
pytest-env==0.8.1
    # via -r requirements.in
pytest-mimesis==1.1.0
    # via -r requirements.in
pytest-randomly==3.12.0
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via docker
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   asgi-lifespan
    #   httpcore
    #   httpx
starlette==0.22.0
    # via fastapi
tomli==2.0.1
    # via
    #   autoflake
    #   black
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via
    #   black
    #   mypy
    #   pydantic
    #   starlette
urllib3==1.26.13
    # via
    #   docker
    #   requests
uvicorn==0.20.0
    # via -r requirements.in
websocket-client==1.4.2
    # via docker
zipp==3.11.0
    # via importlib-metadata
